[61] Adding additional changes, for size reduction on both ancient and active

+++ core/rawdb/freezer.go
@@ -61,7 +61,7 @@ const (
 
 	// freezerBatchLimit is the maximum number of blocks to freeze in one batch
 	// before doing an fsync and deleting it from the key-value store.
-	freezerBatchLimit = 9
+	freezerBatchLimit = 9 // mike
 
 	// freezerTableSize defines the maximum size of freezer data files.
 	freezerTableSize = 2 * 1000 * 1000 * 1000
@@ -70,10 +70,10 @@ const (
 // freezer is a memory mapped append-only database to store immutable chain data
 // into flat files:
 //
-// - The append only nature ensures that disk writes are minimized.
-// - The memory mapping ensures we can max out system memory for caching without
-//   reserving it for go-ethereum. This would also reduce the memory requirements
-//   of Geth, and thus also GC overhead.
+//   - The append only nature ensures that disk writes are minimized.
+//   - The memory mapping ensures we can max out system memory for caching without
+//     reserving it for go-ethereum. This would also reduce the memory requirements
+//     of Geth, and thus also GC overhead.
 type freezer struct {
 	// WARNING: The `frozen` field is accessed atomically. On 32 bit platforms, only
 	// 64-bit aligned fields can be atomic. The struct is guaranteed to be so aligned,
@@ -96,6 +96,8 @@ type freezer struct {
 	quit      chan struct{}
 	wg        sync.WaitGroup
 	closeOnce sync.Once
+
+	offset uint64 // Starting BlockNumber in current freezer
 }
 
 // newFreezer creates a chain freezer that moves ancient chain data into
@@ -103,7 +105,7 @@ type freezer struct {
 //
 // The 'tables' argument defines the data tables. If the value of a map
 // entry is true, snappy compression is disabled for the table.
-func newFreezer(datadir string, namespace string, readonly bool, maxTableSize uint32, tables map[string]bool) (*freezer, error) {
+func newFreezer(datadir string, namespace string, readonly bool, offset uint64, maxTableSize uint32, tables map[string]bool) (*freezer, error) {
 	// Create the initial freezer object
 	var (
 		readMeter  = metrics.NewRegisteredMeter(namespace+"ancient/read", nil)
@@ -131,6 +133,7 @@ func newFreezer(datadir string, namespace string, readonly bool, maxTableSize ui
 		instanceLock: lock,
 		trigger:      make(chan chan struct{}),
 		quit:         make(chan struct{}),
+		offset:       offset,
 	}
 
 	// Create the tables.
@@ -175,10 +178,14 @@ func newFreezer(datadir string, namespace string, readonly bool, maxTableSize ui
 		return nil, err
 	}
 
+	// Some blocks in ancientDB may have already been frozen and been pruned, so adding the offset to
+	// reprensent the absolute number of blocks already frozen.
+	freezer.frozen += offset
+
 	// Create the write batch.
 	freezer.writeBatch = newFreezerBatch(freezer)
 
-	log.Info("Opened ancient database", "database", datadir, "readonly", readonly)
+	log.Info("Opened ancient database", "database", datadir, "readonly", readonly, "frozen", freezer.frozen, "offset", freezer.offset)
 	return freezer, nil
 }
 
@@ -211,7 +218,7 @@ func (f *freezer) Close() error {
 // in the freezer.
 func (f *freezer) HasAncient(kind string, number uint64) (bool, error) {
 	if table := f.tables[kind]; table != nil {
-		return table.has(number), nil
+		return table.has(number - f.offset), nil
 	}
 	return false, nil
 }
@@ -219,16 +226,16 @@ func (f *freezer) HasAncient(kind string, number uint64) (bool, error) {
 // Ancient retrieves an ancient binary blob from the append-only immutable files.
 func (f *freezer) Ancient(kind string, number uint64) ([]byte, error) {
 	if table := f.tables[kind]; table != nil {
-		return table.Retrieve(number)
+		return table.Retrieve(number - f.offset)
 	}
 	return nil, errUnknownTable
 }
 
 // AncientRange retrieves multiple items in sequence, starting from the index 'start'.
 // It will return
-//  - at most 'max' items,
-//  - at least 1 item (even if exceeding the maxByteSize), but will otherwise
-//   return as many items as fit into maxByteSize.
+//   - at most 'max' items,
+//   - at least 1 item (even if exceeding the maxByteSize), but will otherwise
+//     return as many items as fit into maxByteSize.
 func (f *freezer) AncientRange(kind string, start, count, maxBytes uint64) ([][]byte, error) {
 	if table := f.tables[kind]; table != nil {
 		return table.RetrieveItems(start, count, maxBytes)
@@ -239,6 +246,16 @@ func (f *freezer) AncientRange(kind string, start, count, maxBytes uint64) ([][]
 // Ancients returns the length of the frozen items.
 func (f *freezer) Ancients() (uint64, error) {
 	return atomic.LoadUint64(&f.frozen), nil
+}
+
+// ItemAmountInAncient returns the actual length of current ancientDB.
+func (f *freezer) ItemAmountInAncient() (uint64, error) {
+	return atomic.LoadUint64(&f.frozen) - atomic.LoadUint64(&f.offset), nil
+}
+
+// AncientOffSet returns the offset of current ancientDB.
+func (f *freezer) AncientOffSet() uint64 {
+	return atomic.LoadUint64(&f.offset)
 }
 
 // Tail returns the number of first stored item in the freezer.
@@ -320,7 +337,7 @@ func (f *freezer) TruncateHead(items uint64) error {
 		return nil
 	}
 	for _, table := range f.tables {
-		if err := table.truncateHead(items); err != nil {
+		if err := table.truncateHead(items - f.offset); err != nil {
 			return err
 		}
 	}
@@ -339,8 +356,9 @@ func (f *freezer) TruncateTail(tail uint64) error {
 	if atomic.LoadUint64(&f.tail) >= tail {
 		return nil
 	}
+
 	for _, table := range f.tables {
-		if err := table.truncateTail(tail); err != nil {
+		if err := table.truncateTail(tail - f.offset); err != nil {
 			return err
 		}
 	}
@@ -424,8 +442,9 @@ func (f *freezer) repair() error {
 // This functionality is deliberately broken off from block importing to avoid
 // incurring additional data shuffling delays on block propagation.
 func (f *freezer) freeze(db ethdb.KeyValueStore) {
-	nfdb := &nofreezedb{KeyValueStore: db}
+	isDeleteActive := true
 
+	nfdb := &nofreezedb{KeyValueStore: db}
 	var (
 		backoff   bool
 		triggered chan struct{} // Used in tests
@@ -495,6 +514,8 @@ func (f *freezer) freeze(db ethdb.KeyValueStore) {
 		if limit-first > freezerBatchLimit {
 			limit = first + freezerBatchLimit
 		}
+		
+		log.Info("[ucc] core rawdb freezer --- freeze -- ", "first", first, "freezerBatchLimit", freezerBatchLimit, "limit", limit)
 		ancients, err := f.freezeRange(nfdb, first, limit)
 		if err != nil {
 			log.Error("Error in block freeze operation", "err", err)
@@ -507,46 +528,88 @@ func (f *freezer) freeze(db ethdb.KeyValueStore) {
 			log.Crit("Failed to flush frozen tables", "err", err)
 		}
 
-		// Wipe out older data from the active database
-		batchActive := db.NewBatch()
-		timePrune := time.Now()
+		if isDeleteActive {
+			// Wipe out older data from the active database
+			batchActive := db.NewBatch()
+			timePrune := time.Now()
+			timeCompactIteration := time.Now()
 
-		var activeHashes []common.Hash
-		var activeNumber uint64
+			var activeHashes []common.Hash
+			var activeNumber, itDeleteActive uint64
 
-		headNumber := ReadHeaderNumber(nfdb, hash)
+			headNumber := ReadHeaderNumber(nfdb, hash)
 
-		if *headNumber < 552120 {
-			log.Info("[ucc] active data still less than magic number ---- ")
-		} else {
-			endBlockToDelete := *headNumber - 552120
-			log.Info("[ucc] active data delete --- ", "start", 0, "end", endBlockToDelete)
+			if *headNumber < 552120 {
+				log.Info("[ucc] active data still less than magic number ---- ")
+			} else {
+				endBlockToDelete := *headNumber - 552120
+				log.Info("[ucc] active data delete --- ", "start", 0, "end", endBlockToDelete)
 
-			for activeNumber = 0; activeNumber < endBlockToDelete; activeNumber++ {
-				// Always keep the genesis block in active database
-				if activeNumber != 0 {
-					activeHashes = ReadAllHashes(db, activeNumber)
-					for _, hash := range activeHashes {
-						DeleteBlock(batchActive, hash, activeNumber)
+				for itDeleteActive = uint64(0); itDeleteActive < endBlockToDelete; itDeleteActive += 10000 {
+					timeCompactIteration = time.Now()
+
+					batchActive = db.NewBatch()
+
+					for activeNumber = 0; activeNumber < endBlockToDelete; activeNumber++ {
+						// Always keep the genesis block in active database
+						if activeNumber != 0 {
+							activeHashes = ReadAllHashes(db, activeNumber)
+							for _, hash := range activeHashes {
+								DeleteBlock(batchActive, hash, activeNumber)
+							}
+						}
+					}
+				
+					if err := batchActive.Write(); err != nil {
+						log.Crit("[ucc] active data delete failed ---- ", "err", err)
+					}
+
+					log.Info("[ucc] active data delete iteration --- ", "it", itDeleteActive, "elapsed", common.PrettyDuration(time.Since(timeCompactIteration)))
+						
+					batchActive.Reset()
+				}
+
+				log.Info("[ucc] active data delete completed --- ", "elapsed", common.PrettyDuration(time.Since(timePrune)))
+
+					
+				// --- compaction
+				cstart := time.Now()
+				for bCompact := 0x00; bCompact <= 0xf0; bCompact += 0x10 {
+					var (
+						startCompact = []byte{byte(bCompact)}
+						endCompact   = []byte{byte(bCompact + 0x10)}
+					)
+					if bCompact == 0xf0 {
+						endCompact = nil
+					}
+					log.Info("[ucc] Compacting database ---- ", "bCompact", bCompact, "range", fmt.Sprintf("%#x-%#x", startCompact, endCompact), "elapsed", common.PrettyDuration(time.Since(cstart)))
+					if err := db.Compact(startCompact, endCompact); err != nil {
+						log.Error("[ucc] Database compaction failed ---- ", "error", err)
 					}
 				}
-			}
-	
-			if err := batchActive.Write(); err != nil {
-				log.Crit("[ucc] active data delete failed ---- ", "err", err)
+				log.Info("[ucc] Database compaction finished ---- ", "elapsed", common.PrettyDuration(time.Since(cstart)))
+
+				log.Info("[ucc] State pruning successful ---- ", "elapsed", common.PrettyDuration(time.Since(timePrune)))
+				// --- end compaction
 			}
 
-			log.Info("[ucc] active data delete duration --- ", "elapsed", common.PrettyDuration(time.Since(timePrune)))
-			
-			batchActive.Reset()
 		}
+
 		batch := db.NewBatch()
 		for i := 0; i < len(ancients); i++ {
 			// Always keep the genesis block in active database
 			if first+uint64(i) != 0 {
 				log.Info("[ucc] core rawdb freezer --- freeze -- deleting data # -- ancient ", "first", first,  "i", uint64(i), "first+uint64(i", first+uint64(i))
-				DeleteBlockWithoutNumber(batch, ancients[i], first+uint64(i))
+
+				// rawdb.DeleteCanonicalHash(batch, block.NumberU64())
+				// rawdb.DeleteBlockWithoutNumber(batch, block.Hash(), block.NumberU64())
+	
+				// func DeleteBlockWithoutNumber(db ethdb.KeyValueWriter, hash common.Hash, number uint64) {
+
+				DeleteBlockWithoutNumber(batch, ancients[i], first+uint64(i)) 
 				DeleteCanonicalHash(batch, first+uint64(i))
+			} else {
+				log.Info("[ucc] core rawdb freezer --- freeze -- deleting data # -- keeping genesis on # ", uint64(i))
 			}
 		}
 		if err := batch.Write(); err != nil {
@@ -613,6 +676,7 @@ func (f *freezer) freeze(db ethdb.KeyValueStore) {
 			context = append(context, []interface{}{"hash", ancients[n-1]}...)
 		}
 		log.Info("Deep froze chain segment", context...)
+		log.Info("[ucc] checking thrashing with tiny writes --- ", "f.frozen-first", f.frozen-first, "freezerBatchLimit", freezerBatchLimit, "f.frozen-first < freezerBatchLimit", f.frozen-first < freezerBatchLimit)
 
 		// Avoid database thrashing with tiny writes
 		if f.frozen-first < freezerBatchLimit {
@@ -623,6 +687,7 @@ func (f *freezer) freeze(db ethdb.KeyValueStore) {
 
 func (f *freezer) freezeRange(nfdb *nofreezedb, number, limit uint64) (hashes []common.Hash, err error) {
 	hashes = make([]common.Hash, 0, limit-number)
+	log.Info("[ucc] freezeRange called --- ", "number", number, "limit", limit, "hashes", hashes)
 
 	_, err = f.ModifyAncients(func(op ethdb.AncientWriteOp) error {
 		for ; number <= limit; number++ {
@@ -736,7 +801,7 @@ func (f *freezer) MigrateTable(kind string, convert convertLegacyFn) error {
 		return err
 	}
 	var (
-		batch  = newTable.newBatch()
+		batch  = newTable.newBatch(f.offset)
 		out    []byte
 		start  = time.Now()
 		logged = time.Now()
